# StudentPortal.API/bruno/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "StudentPortal API",
    "description": "PRN232 Lab 1 – RESTful API for managing Students, Courses, Subjects and Semesters.",
    "contact": {
      "name": "PRN232 Student",
      "email": "student@fpt.edu.vn"
    },
    "version": "v1"
  },
  "paths": {
    "/api/courses": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Get all courses with pagination, search, sorting, select, and expand options.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortby",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortdescending",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "$select",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "$expand",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a paged list of courses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectPagedResponseApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Create a new course.",
        "requestBody": {
          "description": "Course data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or semester not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{id}": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Get a single course by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Courses"
        ],
        "summary": "Update an existing course.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Updated data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "summary": "Delete a course by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CourseResponse": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int32"
          },
          "courseName": {
            "type": "string",
            "nullable": true
          },
          "semesterId": {
            "type": "integer",
            "format": "int32"
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubjectResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CourseResponseApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/CourseResponse"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
      "CreateCourseRequest": {
        "required": [
          "courseName",
          "semesterId"
        ],
        "type": "object",
        "properties": {
          "courseName": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string",
            "example": "PRN232 – ASP.NET Web API Development"
          },
          "semesterId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "example": 1
          }
        },
        "additionalProperties": false,
        "description": "Payload for creating a new Course."
      },
...
      "ObjectApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
      "ObjectPagedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "description": "Pagination metadata wrapper."
      },
      "ObjectPagedResponseApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/ObjectPagedResponse"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
...
      "SubjectResponse": {
        "type": "object",
        "properties": {
          "subjectId": {
            "type": "integer",
            "format": "int32"
          },
          "subjectCode": {
            "type": "string",
            "nullable": true
          },
          "subjectName": {
            "type": "string",
            "nullable": true
          },
          "credit": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Base subject response — no navigation properties."
      },
...
      "UpdateCourseRequest": {
        "required": [
          "courseName",
          "semesterId"
        ],
        "type": "object",
        "properties": {
          "courseName": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string",
            "example": "PRN232 – ASP.NET Web API Development"
          },
          "semesterId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "example": 1
          }
        },
        "additionalProperties": false,
        "description": "Payload for updating an existing Course."
      },
...
    }
  }
}